backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
scikit-learn==1.8.0
pandas==2.2.3
joblib==1.4.2
numpy==1.26.4
bcrypt==4.0.1
psycopg2-binary==2.9.9
PySastrawi==1.2.0
